refacto companies end
